--- Api data Handling --- UI setup --- Dashboard Handling

+++ app/src/main/java/com/app/weatherapp/models/currentTempModels/Clouds.kt
@@ -1,5 +1,5 @@
 package com.app.weatherapp.models.currentTempModels
 
 data class Clouds(
-    val all: Int?
+    val all: Double?
 )
