+++ app/src/main/java/com/app/weatherapp/models/currentTempModels/Weather.kt
@@ -3,6 +3,6 @@ package com.app.weatherapp.models.currentTempModels
 data class Weather(
     val description: String?,
     val icon: String?,
-    val id: Int?,
+    val id: Double?,
     val main: String?
 )
