+++ app/src/main/java/com/app/weatherapp/models/currentTempModels/Wind.kt
@@ -1,7 +1,7 @@
 package com.app.weatherapp.models.currentTempModels
 
 data class Wind(
-    val deg: Int?,
+    val deg: Double?,
     val gust: Double?,
     val speed: Double?
 )
