+++ app/src/main/java/com/app/weatherapp/models/forcastDataModels/Data.kt
@@ -6,13 +6,13 @@ import com.app.weatherapp.models.currentTempModels.Wind
 
 data class Data (
     val clouds: Clouds?,
-    val dt: Int?,
+    val dt: Double?,
     val dt_txt: String?,
     val main: Main?,
-    val pop: Int?,
+    val pop: Double?,
     val rain: Rain?,
     val sys: Sys?,
-    val visibility: Int?,
+    val visibility: Double?,
     val weather: List<Weather>?,
     val wind: Wind?
 )
