+++ app/src/main/java/com/app/weatherapp/models/forcastDataModels/ForecastDataResponse.kt
@@ -2,8 +2,8 @@ package com.app.weatherapp.models.forcastDataModels
 
 data class ForecastDataResponse(
     val city: City?,
-    val cnt: Int?,
+    val cnt: Double?,
     val cod: String?,
     val list: List<Data>?,
-    val message: Int?
+    val message: Double?
 )
